fix issue where neovim wasn't autoformatting on save
Autoformatting on save would work after intially entering a buffer but if the buffer was left and reentered, autoformatting wouldn't work. It was because the augroup was being created in the scope of the autoformat_on_save function. Moving it to the top level of the module fixed the issue.